+++ app/src/main/res/layout/fragment_test_render.xml
@@ -12,7 +12,6 @@
         android:layout_height="0dp"
         android:layout_margin="16dp"
         android:background="@drawable/text_input_border"
-        android:contentDescription="testRenderMarkupEditText"
         android:imeOptions="actionSend"
         android:padding="5dp"
         android:textColor="@color/turquoise"
@@ -31,7 +30,6 @@
         android:layout_marginHorizontal="45dp"
         android:layout_marginTop="30dp"
         android:background="@drawable/round_button"
-        android:contentDescription="testRenderTestButton"
         android:fontFamily="@font/proxima_nova_bold"
         android:text="@string/test_render_button"
         android:textColor="@android:color/white"
